requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0  # PostgreSQL (Supabase)
aiomysql==0.2.0  # MySQL (XAMPP Local)
PyMySQL==1.1.0   # MySQL driver alternativo

# HTTP Client
httpx==0.26.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Google AI
google-generativeai==0.3.2

# Utilities
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Monitoring (Optional)
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

